Extracts ContestedArmiesCalculator from CombatResolver
