+++ core/src/main/combat/ContestedArmiesCalculator.kt
@@ -1,0 +1,19 @@
+package combat
+
+import PositiveInt
+
+class ContestedArmiesCalculator(
+    private val attackingArmies: PositiveInt, private val defendingArmies: PositiveInt) {
+    fun getArmies(): PositiveInt {
+        val maxAttackerContestedArmies = attackingArmies - PositiveInt(1)
+        return minOf(
+            maxAttackerContestedArmies,
+            defendingArmies,
+            MAX_CONTESTED_ARMIES
+        )
+    }
+
+    companion object {
+        val MAX_CONTESTED_ARMIES = PositiveInt(3)
+    }
+}
